+++ app/Http/Controllers/TeacherController.php
@@ -6,6 +6,7 @@ use App\Models\User;
 use App\Models\Course;
 use Illuminate\Http\Request;
 use App\Events\UserCreateEvent;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 
 class TeacherController extends Controller
@@ -17,6 +18,10 @@ class TeacherController extends Controller
      */
     public function index()
     {
+        if(!Gate::allows('modif-akun')){
+            abort(403);
+        }
+
         $data = User::whereHas('roles', function ($q){
             $q->where('name', config('enums.roles.teacher') );
         })->latest()->paginate();
@@ -31,6 +36,10 @@ class TeacherController extends Controller
      */
     public function create()
     {
+        if(!Gate::allows('modif-akun')){
+            abort(403);
+        }
+
         return view('pages.teacher.create');
     }
 
@@ -42,6 +51,10 @@ class TeacherController extends Controller
      */
     public function store(Request $request)
     {
+        if(!Gate::allows('modif-akun')){
+            abort(403);
+        }
+
         $fields = $request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -70,7 +83,9 @@ class TeacherController extends Controller
      */
     public function show($id)
     {
-        //
+        if(!Gate::allows('modif-akun')){
+            abort(403);
+        }
     }
 
     /**
@@ -81,6 +96,10 @@ class TeacherController extends Controller
      */
     public function edit($id)
     {
+        if(!Gate::allows('modif-akun')){
+            abort(403);
+        }
+
         $teacher= User::findOrFail($id);
 
         return view('pages.teacher.edit', compact('teacher'));
@@ -95,6 +114,10 @@ class TeacherController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!Gate::allows('modif-akun')){
+            abort(403);
+        }
+
         $fields = $request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -115,6 +138,10 @@ class TeacherController extends Controller
      */
     public function destroy($id)
     {
+        if(!Gate::allows('modif-akun')){
+            abort(403);
+        }
+
         $teacher= User::findOrFail($id);
 
         $teacher->delete();
